# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Buddy Script API",
        "description": "API Documentation for the Buddy Script Backend API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server Host"
        }
    ],
    "paths": {
...
        "/api/posts/{postId}/react": {
            "post": {
                "tags": [
                    "Reactions"
                ],
                "summary": "React to Post",
                "description": "Add, update, or remove reaction on a post. Sending the exact same reaction_type will toggle it off (delete it).",
                "operationId": "40d3611d36967546f0275bae0f5cf1ca",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "Target Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reaction_type"
                                ],
                                "properties": {
                                    "reaction_type": {
                                        "type": "string",
                                        "example": "like",
                                        "enum": [
                                            "like",
                                            "love",
                                            "haha",
                                            "wow",
                                            "sad",
                                            "angry"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reaction registered or updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reaction registered"
                                        },
                                        "data": {
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/Reaction"
                                                }
                                            ],
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Post not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/comments/{commentId}/react": {
            "post": {
                "tags": [
                    "Reactions"
                ],
                "summary": "React to Comment",
                "description": "Add, update, or remove reaction on a comment/reply. Sending the exact same reaction_type toggles it off.",
                "operationId": "01f0d51cc8e30b8f88267c24db72be75",
                "parameters": [
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "Target Comment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reaction_type"
                                ],
                                "properties": {
                                    "reaction_type": {
                                        "type": "string",
                                        "example": "love",
                                        "enum": [
                                            "like",
                                            "love",
                                            "haha",
                                            "wow",
                                            "sad",
                                            "angry"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reaction updated or toggled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reaction registered"
                                        },
                                        "data": {
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/Reaction"
                                                }
                                            ],
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/posts/{postId}/reactions": {
            "get": {
                "tags": [
                    "Reactions"
                ],
                "summary": "Get Post Reaction List",
                "description": "Fetch a paginated list of users who reacted to a post. Optionally filter by reaction_type.",
                "operationId": "859d19ead3d2ff864c06457f39274c8c",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "Target Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "reaction_type",
                        "in": "query",
                        "description": "Filter by type (like, love, etc.)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reaction list retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ReactionUser"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found"
                    }
                }
            }
        },
        "/api/comments/{commentId}/reactions": {
            "get": {
                "tags": [
                    "Reactions"
                ],
                "summary": "Get Comment Reaction List",
                "description": "Fetch a paginated list of users who reacted to a specific comment/reply.",
                "operationId": "1b4ad626539381b535fa8459d23b9941",
                "parameters": [
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "Target Comment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reaction list retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ReactionUser"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "first_name": {
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-12T13:22:00Z"
                    }
                },
                "type": "object"
            },
...
            "Reaction": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "reaction_type": {
                        "type": "string",
                        "example": "like"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-12T13:22:00Z"
                    }
                },
                "type": "object"
            },
            "ReactionUser": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "first_name": {
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "reaction_type": {
                        "type": "string",
                        "example": "like"
                    },
                    "reacted_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-12T13:22:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}